+++ rustyhack_client/src/screens/bottom_text_window.rs
@@ -3,8 +3,8 @@ use console_engine::screen::Screen;
 
 pub(crate) fn draw(status_messages: &[String]) -> Screen {
     let mut screen = Screen::new(
-        client_consts::CONSOLE_WIDTH,
-        client_consts::CONSOLE_HEIGHT - client_consts::VIEWPORT_HEIGHT,
+        client_consts::CONSOLE_WIDTH as u32,
+        (client_consts::CONSOLE_HEIGHT - client_consts::VIEWPORT_HEIGHT) as u32,
     );
     if !status_messages.is_empty() {
         for (count, message) in status_messages.iter().rev().enumerate() {
