+++ rustyhack_lib/src/background_map/character_map.rs
@@ -2,7 +2,7 @@ use crate::background_map::tiles::door::Door;
 use crate::background_map::tiles::wall::Wall;
 use crate::background_map::tiles::{OpenState, Tile};
 
-pub fn map_character_to_tile(x: i32, y: i32, character: char) -> Tile {
+pub fn map_character_to_tile(x: isize, y: isize, character: char) -> Tile {
     match character {
         '\n' => Tile::NewLine,
         '\r' => Tile::CarriageReturn,
