+++ rustyhack_lib/src/background_map/tiles/wall.rs
@@ -9,7 +9,7 @@ pub struct Wall {
 }
 
 impl Wall {
-    pub fn new(x: i32, y: i32, character_icon: char) -> Wall {
+    pub fn new(x: isize, y: isize, character_icon: char) -> Wall {
         Wall {
             tile_position: TilePosition { x, y },
             character_icon,
